refactor(cleanup): Extracted clickhouse helpers
