+++ src/zod/query.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import { QUERY_FIELD_NAMES } from "../interface/storage/Storage";
 import { Operator, AggregationType, LogicalOperator } from "../gen/query/v1/query_pb";
+import { createFilterGroupSchema } from "./internals";
 
 const OPERATOR_MAP = {
   [Operator.EQ]: "EQ",
@@ -32,32 +33,7 @@ const filterConditionSchema = z.object({
   value: z.string(),
 });
 
-interface FilterGroupOutput {
-  logical: "AND" | "OR";
-  conditions: Array<z.output<typeof filterConditionSchema>>;
-  groups: FilterGroupOutput[];
-}
-
-const filterGroupSchema: z.ZodType<FilterGroupOutput> = z.lazy(() =>
-  z
-    .object({
-      logical: z
-        .number()
-        .int()
-        .min(0)
-        .max(2)
-        .transform(
-          (v) => LOGICAL_MAP[v as keyof typeof LOGICAL_MAP]
-        ),
-      conditionsList: z.array(filterConditionSchema).default([]),
-      groupsList: z.array(filterGroupSchema).default([]),
-    })
-    .transform((v) => ({
-      logical: v.logical,
-      conditions: v.conditionsList,
-      groups: v.groupsList,
-    }))
-);
+const filterGroupSchema = createFilterGroupSchema(filterConditionSchema, LOGICAL_MAP);
 
 const aggregationSchema = z.object({
   type: z
